server/requirements.txt
aiohttp
aiosignal
alembic
alembic-autogen-check
algorithm
altair
aniso8601
antlr4-python3-runtime
anyio
appdirs
appnope
arize-phoenix
asgiref
asttokens
astunparse
async-timeout
attrbox
attrs
audioread
axios
backcall
backports.zoneinfo
beautifulsoup4
black
bleach
blinker
blis
braceexpand
cachetools
catalogue
certifi
cffi
cfgv
chardet
charset-normalizer
Chroma
click
cloudpathlib
colorama
colorgram.py
confection
Contacts
contexttimer
contourpy
cryptography
cycler
cymem
dataclasses-json
decorator
decord
dependency-injection
Deprecated
distlib
distro
docker
docopt
einops
et-xmlfile
exceptiongroup
executing
fairscale
faiss-cpu
fastapi
filelock
Flask
Flask-Cors
fonttools
frozenlist
fsspec
ftfy
fuelcell
geographiclib
geopy
gitdb
GitPython
googlemaps
graphene
graphene-sqlalchemy
graphql-core
graphql-relay
greenlet
h11
httpcore
httptools
httpx
huggingface-hub
identify
idna
imageio
imessage-reader
importlib-metadata
importlib-resources
imports
iniconfig
iopath
ipython
isort
itsdangerous
jedi
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kaggle
kiwisolver
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langcodes
langsmith
lazy_loader
librosa
llvmlite
lxml
Mako
maps
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mpmath
msgpack
multidict
murmurhash
mypy-extensions
networkx
nodeenv
numba
numpy
omegaconf
openai
opendatasets
openpyxl
optionlab
orjson
packaging
pandas
parso
pathspec
pdfmerge
pexpect
phoenix
pickleshare
Pillow
pipreqs
pkgutil_resolve_name
platformdirs
plotly
pluggy
pooch
portalocker
pre-commit
preshed
promise
prompt-toolkit
protobuf
psycopg2-binary
ptyprocess
pure-eval
py-iMessage
pyarrow
pycocoevalcap
pycocotools
pycparser
pydantic
pydeck
Pygments
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-Contacts
pyparsing
pypdf
PyPDF2
PyQt5
PyQt5-Qt5
PyQt5-sip
PySide2
pytest
python-dateutil
python-dotenv
python-editor
python-magic
python-slugify
python-tsp
pytz
pyvespa
PyWavelets
PyYAML
referencing
regex
requests
requests-toolbelt
rich
rpds-py
ruff
Rx
salesforce-lavis
scikit-image
scikit-learn
scipy
sentencepiece
shiboken2
singledispatch
six
smart-open
smmap
sniffio
soundfile
soupsieve
soxr
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
sseclient-py
stack-data
star
starlette
starlette-graphene3
strawberry-graphql
streamlit
supervision
sympy
tabulate
tenacity
text-unidecode
thinc
threadpoolctl
tifffile
tiktoken
timm
tk
together
tokenizers
toml
tomli
toolz
torch
torchvision
tornado
tqdm
traitlets
transformers
tsplib95
typer
typing-inspect
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
uvloop
validators
virtualenv
waitress
wasabi
watchdog
watchfiles
wcwidth
weasel
webdataset
webencodings
websockets
Werkzeug
wrapt
xlwt
yarg
yarl
zipp